CocoaPods - Use Version Tag
The current setup for react-native-fast-image always attempts to pull the source from the `mainline` branch.  I just ran into an issue where I was attempting to install an old version of react-native-fast-image, but instead it fetched the mainline version which was incompatible.  

This change ensures the version which is downloaded from NPM matches the version of the source.